onchain tests involving secondary transfers
