+++ store-api-server/package.json
@@ -78,6 +78,8 @@
     "@nestjs/cli": "^8.2.0",
     "@nestjs/schematics": "^8.0.5",
     "@nestjs/testing": "^8.2.6",
+    "@taquito/signer": "^14.0.0",
+    "@taquito/taquito": "^14.0.0",
     "@types/express": "^4.17.11",
     "@types/node": "^14.14.31",
     "@types/supertest": "^2.0.10",
