+++ store-api-server/script/migrate
@@ -4,7 +4,7 @@ cd $SCRIPT_DIR/..
 
 PEPPERMINT_VERSION=ee538be4d156ffb456107587eb71f14671afb1c7
 TEZPAY_VERSION=7229e80e49a4b697dc634011a935f6c102d8c42a
-QUEPASA_VERSION=1.2.5
+QUEPASA_VERSION=${QUEPASA_VERSION:-'1.2.6'}
 
 ./script/wait-db
